Hide/Show Toolbar (#83)
Start on work to allow toolbar to be hidden/shown when mouse goes over.

Involves having a second label - just a few pixels tall. When the mouse
goes over it - hide it & show the toolbar.

When the mouse leaves the toolbar - hide it & show the label.

Includes a function to ```setToolbarPinned()``` to turn it on/off.

Now needs an icon (drawing pin) to place in toolbar to call the pin
function/show pin status.